+++ src/com/tutorial/main/HUD.java
@@ -5,10 +5,10 @@ import java.awt.Graphics;
 
 public class HUD {
 
-   public static int HEALTH = 100;
+   public static float HEALTH = 100;
    
-   private int greenValue = 255;
-   private int redValue = 0;
+   private float greenValue = 255;
+   private float redValue = 0;
    
    private int score = 0;
    private int level = 1;
@@ -31,10 +31,10 @@ public class HUD {
       g.drawString("Health: " + HEALTH + "%", 15, 10);
       g.setColor(Color.gray);
       g.fillRect(15, 15, 200, 32);
-      g.setColor(new Color(redValue, greenValue, 0));
-      g.fillRect(15, 15, HEALTH * 2, 32);
+      g.setColor(new Color((int) redValue, (int) greenValue, 0));
+      g.fillRect(15, 15, (int)HEALTH * 2, 32);
       g.setColor(Color.white);
-      g.drawRect(15, 15, HEALTH * 2, 32);
+      g.drawRect(15, 15, (int)HEALTH * 2, 32);
       
       g.drawString("Score: " + score, 15, 64);
       g.drawString("Level: " + level, 15, 80);
